シェーダーコンパイル関数と、リファレクション関数に名前空間を使用して、分かり易いように修正
