+++ 3DAction_Project/3DAction_Project/ShaderReflectionUtils.h
@@ -25,14 +25,19 @@
 // ==================================================
 // 関数
 // ==================================================
-// リファレンスしたシェーダー外部情報を読込み
-bool ShaderInfoInput(const char* kShaderInfoPath, std::vector<ShaderInfo>& outShaderInfo);
+namespace ShaderReflectionUtils
+{
+	// リファレンスしたシェーダー外部情報を読込み
+	bool ShaderInfoInput(const char* kShaderInfoPath, std::vector<ShaderInfo>& outShaderInfo);
 
 #if defined(DEBUG) || defined(_DEBUG)
-// デバッグ用関数　デバッグ時のみを行う
+	// デバッグ用関数　デバッグ時のみを行う
 
-// リフレクションした情報をShaderInfo配列に入れる
-bool Reflect(void* blob, size_t blobsize, std::vector<ConstantBufferInfo>& CBInfo, std::vector<InputLayoutInfo>& ILInfo);
-// リファレンスしたシェーダー情報配列を外部ファイルに出力する
-bool ShaderInfoOutput(const char* kShaderInfoPath, std::vector<ShaderInfo>& shaderInfo);
+	// リフレクションした情報をShaderInfo配列に入れる
+	bool Reflect(void* blob, size_t blobsize, std::vector<ConstantBufferInfo>& CBInfo, std::vector<InputLayoutInfo>& ILInfo);
+	// リファレンスしたシェーダー情報配列を外部ファイルに出力する
+	bool ShaderInfoOutput(const char* kShaderInfoPath, std::vector<ShaderInfo>& shaderInfo);
+
 #endif
+
+}
